[Security Solution] render alert description component in document flyout in Discover
