+++ src/platform/plugins/shared/discover/public/context_awareness/profile_providers/security/components/enhanced_alert_event_overview.tsx
@@ -7,8 +7,9 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-import React from 'react';
+import React, { useEffect, useState } from 'react';
 import type { DocViewerComponent } from '@kbn/unified-doc-viewer/types';
+import { useDiscoverServices } from '../../../../hooks/use_discover_services';
 
 /**
  * This component is a placeholder for the new alert/event Overview tab content.
@@ -17,5 +18,36 @@ import type { DocViewerComponent } from '@kbn/unified-doc-viewer/types';
  * The feature flag will remain disabled until we're ready to ship some of the content. The target is to release an MVP by 9.4 then have it fully functional by 9.5.
  */
 export const EnhancedAlertEventOverview: DocViewerComponent = ({ hit }) => {
-  return <></>;
+  const { discoverShared } = useDiscoverServices();
+  const alertFlyoutOverviewTabFeature = discoverShared.features.registry.getById(
+    'security-solution-alert-flyout-overview-tab'
+  );
+
+  const [AlertFlyoutOverviewTab, setAlertFlyoutOverviewTab] = useState<(() => JSX.Element) | null>(
+    null
+  );
+
+  useEffect(() => {
+    let isMounted = true;
+
+    const load = async () => {
+      const render = alertFlyoutOverviewTabFeature?.render;
+      if (!render) return;
+
+      const component = await render(hit);
+      if (!isMounted) return;
+
+      setAlertFlyoutOverviewTab(() => component);
+    };
+
+    load();
+
+    return () => {
+      isMounted = false;
+    };
+  }, [alertFlyoutOverviewTabFeature, hit]);
+
+  if (!AlertFlyoutOverviewTab) return null;
+
+  return <AlertFlyoutOverviewTab />;
 };
