+++ src/platform/plugins/shared/discover_shared/public/index.ts
@@ -23,6 +23,7 @@ export type {
   ObservabilityStreamsFeature,
   SecuritySolutionCellRendererFeature,
   SecuritySolutionAppWrapperFeature,
+  SecuritySolutionAlertFlyoutOverviewTabFeature,
   DiscoverFeature,
   DiscoverFeaturesServiceSetup,
   DiscoverFeaturesServiceStart,
