+++ x-pack/solutions/security/plugins/security_solution/public/flyout/attack_details/components/ai_summary_section.test.tsx
@@ -8,10 +8,9 @@
 import React from 'react';
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
-
 import { AISummarySection } from './ai_summary_section';
 import { useOverviewTabData } from '../hooks/use_overview_tab_data';
-import { useExpandSection } from '../../shared/hooks/use_expand_section';
+import { useExpandSection } from '../../../flyout_v2/shared/hooks/use_expand_section';
 
 jest.mock('@kbn/i18n-react', () => ({
   FormattedMessage: ({ defaultMessage }: { defaultMessage: string }) => <>{defaultMessage}</>,
@@ -21,11 +20,11 @@ jest.mock('../hooks/use_overview_tab_data', () => ({
   useOverviewTabData: jest.fn(),
 }));
 
-jest.mock('../../shared/hooks/use_expand_section', () => ({
+jest.mock('../../../flyout_v2/shared/hooks/use_expand_section', () => ({
   useExpandSection: jest.fn(),
 }));
 
-jest.mock('../../shared/components/expandable_section', () => ({
+jest.mock('../../../flyout_v2/shared/components/expandable_section', () => ({
   ExpandableSection: ({
     title,
     children,
