+++ x-pack/solutions/security/plugins/security_solution/public/flyout/attack_details/components/insights_section.tsx
@@ -7,9 +7,8 @@
 
 import React, { memo } from 'react';
 import { FormattedMessage } from '@kbn/i18n-react';
-
-import { useExpandSection } from '../../shared/hooks/use_expand_section';
-import { ExpandableSection } from '../../shared/components/expandable_section';
+import { useExpandSection } from '../../../flyout_v2/shared/hooks/use_expand_section';
+import { ExpandableSection } from '../../../flyout_v2/shared/components/expandable_section';
 import { FLYOUT_STORAGE_KEYS } from '../constants/local_storage';
 import { INSIGHTS_SECTION_TEST_ID } from '../constants/test_ids';
 import { EntitiesOverview } from './entities_overview';
