+++ x-pack/solutions/security/plugins/security_solution/public/flyout/rule_details/right/test_ids.ts
@@ -5,8 +5,11 @@
  * 2.0.
  */
 
+import {
+  CONTENT_TEST_ID,
+  HEADER_TEST_ID,
+} from '../../../flyout_v2/shared/components/expandable_section';
 import { PREFIX } from '../../shared/test_ids';
-import { CONTENT_TEST_ID, HEADER_TEST_ID } from '../../shared/components/expandable_section';
 
 const RULE_PANEL_TEST_ID = `${PREFIX}RulePanel` as const;
 export const RULE_TITLE_TEST_ID = `${RULE_PANEL_TEST_ID}Title` as const;
