+++ x-pack/solutions/security/plugins/security_solution/public/one_discover/alert_flyout_overview_tab_component/index.tsx
@@ -1,0 +1,14 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import React from 'react';
+import type { DataTableRecord } from '@kbn/discover-utils';
+import { OverviewTab } from '../../flyout_v2/document_details/tabs/overview_tab';
+
+export const getAlertFlyoutOverviewTabComponent = (hit: DataTableRecord) => {
+  return <OverviewTab hit={hit} />;
+};
