+++ x-pack/solutions/security/plugins/security_solution/public/one_discover/index.tsx
@@ -6,3 +6,4 @@
  */
 
 export { getCellRendererForGivenRecord } from './cell_renderers';
+export { getAlertFlyoutOverviewTabComponent } from './alert_flyout_overview_tab_component';
