+++ x-pack/solutions/security/plugins/security_solution/public/plugin.tsx
@@ -21,7 +21,10 @@ import { AppStatus, DEFAULT_APP_CATEGORIES } from '@kbn/core/public';
 import { Storage } from '@kbn/kibana-utils-plugin/public';
 import type { Logger } from '@kbn/logging';
 import { uiMetricService } from '@kbn/cloud-security-posture-common/utils/ui_metrics';
-import type { SecuritySolutionCellRendererFeature } from '@kbn/discover-shared-plugin/public/services/discover_features';
+import type {
+  SecuritySolutionAlertFlyoutOverviewTabFeature,
+  SecuritySolutionCellRendererFeature,
+} from '@kbn/discover-shared-plugin/public/services/discover_features';
 import { ProductFeatureSecurityKey } from '@kbn/security-solution-features/keys';
 import { ProductFeatureAssistantKey } from '@kbn/security-solution-features/src/product_features_keys';
 import { getLazyCloudSecurityPosturePliAuthBlockExtension } from './cloud_security_posture/lazy_cloud_security_posture_pli_auth_block_extension';
@@ -288,6 +291,7 @@ export class Plugin implements IPlugin<PluginSetup, PluginStart, SetupPlugins, S
   public async registerDiscoverSharedFeatures(plugins: SetupPlugins) {
     const { discoverShared } = plugins;
     const discoverFeatureRegistry = discoverShared.features.registry;
+
     const cellRendererFeature: SecuritySolutionCellRendererFeature = {
       id: 'security-solution-cell-renderer',
       getRenderer: async () => {
@@ -295,8 +299,16 @@ export class Plugin implements IPlugin<PluginSetup, PluginStart, SetupPlugins, S
         return getCellRendererForGivenRecord;
       },
     };
-
     discoverFeatureRegistry.register(cellRendererFeature);
+
+    const alertFlyoutOverviewTabFeature: SecuritySolutionAlertFlyoutOverviewTabFeature = {
+      id: 'security-solution-alert-flyout-overview-tab',
+      render: async (hit) => {
+        const { getAlertFlyoutOverviewTabComponent } = await this.getLazyDiscoverSharedDeps();
+        return () => getAlertFlyoutOverviewTabComponent(hit);
+      },
+    };
+    discoverFeatureRegistry.register(alertFlyoutOverviewTabFeature);
   }
 
   public async getLazyDiscoverSharedDeps() {
